# snapshot: delegate reading to arrow/parquet readers