We were previously copying all snapshot part bytes into a slice that was then passed to arrow readers (parquet copies are left untouched, see https://github.com/polarsignals/frostdb/pull/590#issuecomment-1814523067). This was an unnecessary indirection and resulted in extra allocations on startup. Benchmark results on recovery from snapshot only:
```
goos: darwin
goarch: arm64
pkg: github.com/polarsignals/frostdb
          │  benchmain   │              benchnew              │
          │    sec/op    │   sec/op     vs base               │
Replay-12   1680.6m ± 5%   816.0m ± 3%  -51.45% (p=0.002 n=6)

          │   benchmain   │              benchnew               │
          │     B/op      │     B/op      vs base               │
Replay-12   2382.9Mi ± 0%   633.0Mi ± 0%  -73.43% (p=0.002 n=6)

          │  benchmain   │              benchnew              │
          │  allocs/op   │  allocs/op   vs base               │
Replay-12   20.807M ± 0%   5.623M ± 0%  -72.98% (p=0.002 n=6)
```