@asubiotto 2023-11-23
Modified to only avoid the copy for arrow parts