@asubiotto 2023-11-23
Closing since after fixing the bug the improvement is not really noticeable/worth it:
```
goos: darwin
goarch: arm64
pkg: github.com/polarsignals/frostdb
          │ benchmain  │             benchnew              │
          │   sec/op   │   sec/op    vs base               │
Replay-12   1.681 ± 5%   2.545 ± 4%  +51.41% (p=0.002 n=6)

          │  benchmain   │              benchnew              │
          │     B/op     │     B/op      vs base              │
Replay-12   2.327Gi ± 0%   2.278Gi ± 0%  -2.11% (p=0.002 n=6)

          │  benchmain  │             benchnew              │
          │  allocs/op  │  allocs/op   vs base              │
Replay-12   20.81M ± 0%   20.79M ± 0%  -0.10% (p=0.002 n=6)
```