@brancz 2023-11-23
Did you forget to close?